inventory_service/requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
requests==2.31.0
flask-restx==1.3.2
python-dotenv==1.0.0
mysql-connector-python==9.5.0